steam-api: fixed the last message id not being parsed and sent
The last message id, which is needed by poll requests in order to
request new messages, was not being grabbed correctly from the JSON
parser. The JSON parser does not doubly story an integer's value in
string form. This means the plugin needs to handle the actual integer
value for the JSON object, rather than attempting to use strings.

This resolves issues related to messages not being received from polling
operations.

+++ steam/steam-util.h
@@ -31,7 +31,7 @@ void g_slist_free_full(GSList *list, GDestroyNotify free_func);
 gboolean steam_util_json_val(json_value *json, const gchar *name,
                              json_type type, json_value **val);
 
-gboolean steam_util_json_int(json_value *json, const gchar *name, gint *i);
+gboolean steam_util_json_int(json_value *json, const gchar *name, gint64 *i);
 
 gboolean steam_util_json_str(json_value *json, const gchar *name,
                              const gchar **str);
